data/pipeline.py
"""
Data pipeline stub: shows where to integrate TRDG (TextRecognitionDataGenerator)
and the HuggingFace dataset loader.
"""
import os
import json
import numpy as np
import pandas as pd
from PIL import Image
from transformers import AutoTokenizer, AutoModelForCausalLM
from trdg import TextRecognitionDataGenerator

def synthesize_with_trdg(text, out_path):
    # Use TRDG externally to create image for 'text'
    # Example (to run separately): python3 -m trdg.run --text "..." --output ...
    pass

{
    "cells": [
        {
            "cell_type": "markdown",
            "id": "#VSC-ac09160d",
            "metadata": {
                "language": "markdown"
            },
            "source": [
                "**Dataset Analysis**"
            ]
        },
        {
            "cell_type": "code",
            "id": "#VSC-1c90e891",
            "metadata": {
                "language": "python"
            },
            "source": [
                ""
            ]
        }
    ]
}